move queryLockStatus to new teacherPanelHelpers.js
